+++ src/test/resources/model/modelCastsValueProperty.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App {
+    class UserModel extends \Hyperf\Database\Model\Model
+    {
+        protected $table = 'users';
+
+        protected array $casts = ['created_at' => '<caret>'];
+    }
+}