+++ src/test/resources/model/modelFillablePropertyReference.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App {
+    class UserModel extends \Hyperf\Database\Model\Model
+    {
+        protected $table = 'users';
+
+        protected array $fillable = ['id', 'email<caret>'];
+    }
+}